moved some functions from TorrentController into TorrentCRUDController
